Fix user -> Only one login page for all -> chain provider - Makes user's login great again \o/
