+++ src/AppBundle/Controller/Administration/AdminController.php
@@ -7,27 +7,8 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use AppBundle\Controller\Administration\UserManagerInterface;
 
-class AdminController extends Controller implements UserManagerInterface
+class AdminController extends Controller
 {
-    public function loginAction(Request $request)
-    {
-        $form = $this->createForm(SignInType::class);
-
-        $authenticationUtils = $this->get('security.authentication_utils');
-        $error = $authenticationUtils->getLastAuthenticationError();
-        $lastUsername = $authenticationUtils->getLastUsername();
-
-        return $this->render('@App/Admin/signIn.html.twig', [
-            'lastUsername' => $lastUsername,
-            'error' => $error,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    public function logoutAction(Request $request)
-    {
-    }
-
     public function indexAction(Request $request)
     {
 		return $this->render('AppBundle:Admin:index.html.twig');
