+++ src/AppBundle/Controller/Administration/CampaignAdminController.php
@@ -5,20 +5,10 @@ namespace AppBundle\Controller;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 
-class CampaignAdminController extends Controller implements UserManagerInterface
+class CampaignAdminController extends Controller
 {
     public function indexAction()
     {
-        
-    }
 
-    public function loginAction(Request $request)
-    {
-        // TODO: Implement loginAction() method.
-    }
-
-    public function logoutAction(Request $request)
-    {
-        // TODO: Implement logoutAction() method.
     }
 }
