+++ src/AppBundle/Controller/Administration/JurorController.php
@@ -6,20 +6,9 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use AppBundle\Controller\Administration\UserManagerInterface;
 
-class JurorController extends Controller implements UserManagerInterface
+class JurorController extends Controller
 {
     public function indexAction()
     {
-
-    }
-
-    public function loginAction(Request $request)
-    {
-        // TODO: Implement loginAction() method.
-    }
-
-    public function logoutAction(Request $request)
-    {
-        // TODO: Implement logoutAction() method.
     }
 }
