+++ src/AppBundle/Resources/views/Admin/index.html.twig
@@ -13,5 +13,4 @@
             </div>
         </div>
     </div>
-
-    {% endblock %}
+{% endblock %}
