+++ src/AppBundle/Resources/views/Admin/signIn.html.twig
@@ -1,34 +1,34 @@
 {% extends 'AppBundle:Default:layout.html.twig' %}
 
 {% block body %}
-<div class="admin page-wrap">
-    <div class="wc-title-banner">
-        <h2>Login</h2>
-    </div>
+    <div class="admin page-wrap">
+        <div class="wc-title-banner">
+            <h2>Login</h2>
+        </div>
 
-    <div class="wc-container">
-        <div class="col-sm-12">
-            <div class="wc-admin-block">
-    {{ error }}
+        <div class="wc-container">
+            <div class="col-sm-12">
+                <div class="wc-admin-block">
+                    {{ error }}
 
-    <form action="{{ path('public.admin.login') }}" method="post">
-        <label for="username">Username:</label>
-        <input type="text" id="username" name="_username" value="{{ lastUsername }}" />
+                    <form action="{{ path('public.login') }}" method="post">
+                        <label for="username">Username:</label>
+                        <input type="text" id="username" name="_username" value="{{ lastUsername }}"/>
 
-        <label for="password">Password:</label>
-        <input type="password" id="password" name="_password" />
+                        <label for="password">Password:</label>
+                        <input type="password" id="password" name="_password"/>
 
-        {#
-            If you want to control the URL the user
-            is redirected to on success (more details below)
-            <input type="hidden" name="_target_path" value="/account" />
-        #}
+                        {#
+                            If you want to control the URL the user
+                            is redirected to on success (more details below)
+                            <input type="hidden" name="_target_path" value="/account" />
+                        #}
 
-        <button type="submit">login</button>
-    </form>
+                        <button type="submit">login</button>
+                    </form>
+                </div>
             </div>
         </div>
     </div>
-</div>
 {% endblock %}
