+++ src/AppBundle/Resources/views/Default/layout.html.twig
@@ -25,11 +25,11 @@
                 </a>
             </li>
             <li>
-                <a href="{{ path('public.admin.login') }}">
-                    <div class="navIcon">
-                        <i class="fa fa-lock fa-3x navI"></i>
-                    </div>
-                </a>
+                {#<a href="{{ path('public.admin.login') }}">#}
+                    {#<div class="navIcon">#}
+                        {#<i class="fa fa-lock fa-3x navI"></i>#}
+                    {#</div>#}
+                {#</a>#}
             </li>
         </ul>
     </nav>
